+++ sdk/src/main/java/pt/ua/dicoogle/sdk/settings/server/ServerSettings.java
@@ -72,6 +72,8 @@ public interface ServerSettings extends ServerSettingsReader {
         void setDefaultStorage(List<String> storages);
 
         void setNodeName(String nodeName);
+
+        void setCallShutdown(boolean shutdown);
     }
 
     @JsonGetter("dicom-services")