+++ webcore/examples/react-todo/webpack.prod.js
@@ -1,10 +1,6 @@
-const merge = require('webpack-merge');
+const {merge} = require('webpack-merge');
 const common = require('./webpack.common.js');
-const MinifyPlugin = require("babel-minify-webpack-plugin");
 
 module.exports = merge(common, {
   mode: 'production',
-  plugins: [
-    new MinifyPlugin()
-  ]
 });